github: Strip max-age headers before consulting cache
The presence of the max-age headers forces the cache library to use
time-based caching and never send a conditional request with the
cached ETag.  We can add a hook to drop these headers before sending
the response to the caching layer, which will force it to always
use ETag and a conditional request when determing cache freshness.

Resolves-Issue: BonnyCI/projman#229

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>