Fix parsing of github PR url
We parse the owner, project and PR number of a github PR out of the
PR URL.  When using github enterprise, the path preceding
$org/$repo/pulls/$number in the URL may contain some extra stuff which breaks
the parsing here. This fixes the parsing to parse that info from the end of
the string regardless.

This has been submitted to zuul v3 upstream @ https://review.openstack.org/476286

Signed-off-by: Adam Gandelman <adamg@ubuntu.com>